Add music like

+++ src/main/java/com/devgd/melonclone/domain/artist/application/ArtistService.java
@@ -67,11 +67,7 @@ public class ArtistService {
 	}
 
 	public void changeLike(Integer artistId, Integer userId) {
-		System.out.println("artist "+artistId + ", user "+userId);
-		
-		if (artistDao.isLike(artistId, userId))
-			artistDao.removeLike(artistId, userId);
-		else
-			artistDao.addLike(artistId, userId);
+		if (artistDao.isLike(artistId, userId)) artistDao.removeLike(artistId, userId);
+		else artistDao.addLike(artistId, userId);
 	}
 }
